requirements.txt
transformers==4.28.1
pinecone-client==2.2.1
llama_index==0.5.25
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
PyDrive==1.3.1
requests==2.30.0
boto3==1.26.137
tiktoken==0.4.0